(SERVER-862) Added an unregister/lock test and cosmetic test updates
This commit adds one more test case for validating that a lock can be
acquired from the JRubyPool when the last remaining registered instance
has been borrowed before the lock was requested and that borrowed
instance is later unregistered.  A few other cosmetic updates to some
tests are also included.