(SERVER-537) Break flush-pool! into smaller functions
The new shutdown behavior substantially complicated the existing
flush-pool! function, so this commit breaks it down into three new
functions and updates the callers to use the new entry points. The
common functionality has been broken out into swap-and-drain-pool! and
flush-pool! itself now has two entry functions, one for repopulating the
pool (flush-and-repopulate-pool!), and another for readying for shutdown
(flush-pool-for-shutdown!).