(SERVER-862) Readded JRuby waits to a couple of lock tests
This commit readds wait-for-jrubies calls to a couple of
`jruby-locking-tests`.  Those had erroneously been removed in a prior
commit in the PR.  The waits need to be there to ensure that the JRuby
pool is populated with an instance before the test tries to exercise
further code that borrows instances.