(SERVER-537) Roll back test-error-during-init
This commit rolls back the test-error-during-init to a previous
iteration, while removing the logged? test, as it now has more than one
match, which fails the test. It also adds two with-test-logger calls to the
test-error-during-init to try to hide a stacktrace that the test
intentionally induces. This is mostly successful, but occasionally a
stacktrace still appears.