(SERVER-1103) Fix borrow-after-borrow lock tests
This commit includes fixes for timing-related issues in the
`pool-can-borrow-after-borrow-interrupted-during-lock` and
`pool-can-borrow-after-borrow-timed-out-during-lock` tests.  In both
cases, there previously was a race condition where a borrow on a
separate thread could succeed before a lock() call was made, causing the
lock() call to block indefinitely because the pool was indefinitely not
full.  The production code was behaving properly but the tests were
designed with the expectation that the lock would occur before the
any borrows were attempted.  Both tests were changed to wait to perform
the borrow until the lock is taken.