(SERVER-1071) Make legacy routes tolerant of multi-server config
Prior to this commit, the legacy routes services made some
hard-coded assumptions about the master's route configuration that
made it incompatible with the recommended webrouting configuration
style for multi-webserver configurations.  This commit adds a few
more checks to detect that case, and make the services compatible
with the multi-server configuration style.  This is necessary in
order for us to be able to run PE Puppet Server in-process with
other PE web apps.

Conflicts:
	test/integration/puppetlabs/services/legacy_routes/legacy_routes_test.clj