(SERVER-1103) Move try/catch wrapper in test-error-during-init test
This commit moves the try/catch wrapper looking for an exception to be
thrown in the `test-error-during-init` test to be wrapped around the
whole `with-app-with-config` call rather than just the inner `run-app`
call.  The exception could otherwise be thrown before the `run-app` call
is reached if the `create-pool-instance!` call in the test were to occur
quickly enough during initialization.