(SERVER-1103) Added one more `wait-for-jrubies` call to jruby lock test
This commit adds another `wait-for-rubies` call to a test which borrows
a jruby puppet instance.  Without this call, it is possible that a slow
initialization of the jruby pool will cause the test's borrow attempt to
fail because of the relatively short default borrow timeout that the
test uses, 1 millisecond.