(SERVER-1094) Move versioned code logic into core namespace
In general it's a best practice to keep TK service definitions
as minimal as humanly possible.  This has several advantages:

* Makes it possible to write unit tests against functions in
  `core` namespace
* Keeps `core` logic separated from TK namespaces, so there is
  no concrete dependency between the business logic and the
  service framework
* Makes debugging much easier because you're not modifying code
  inside of a macro

This commit moves the business logic from the versioned code
service into a new `core` namespace for the reasons expressed
above.

It also moves the various scripts that are used by the tests
to namespaces in `dev-resources` that correspond to the namespaces
of the tests that consume them.