(SERVER-876) Only call wrap-params-for-jruby once
In a previous commit, wrap-params-for-jruby was called twice for catalog
requests. This meant that the body of the request would be consumed by
the first call (via body-for-jruby), and then the body would be gone for
the second call. This commit avoids that problem by refactoring
root-routes and the wrap-with-code-id to call wrap-params-for-jruby only
once in all routes. It also adds a test for the failing case where
catalog POST requests had no body.