(SERVER-876) Update filebucket tests to include a body
The file_bucket_file tests previously did not include a body, but a
recent change added the wrap-params-with-jruby to the request handler,
which calls a slurp on the body in some cases. This commit adds a body
for both file_bucket_file tests to ensure that there is a body to
slurp/read. This is consistent with the Puppet API docs
https://github.com/puppetlabs/puppet/blob/master/api/docs/http_file_bucket_file.md.