(SERVER-876) Compute code_id when holding a JRuby
This commit refactors the code which computes the code_id to only do so
while holding a JRuby instance.  This ensures that other code which may
result in the code_id changing (such as File Sync) and which requires
use of the JRubyPool's lock cannot be run in between the time when the
code_id is computed and when the catalog is compiled.