(SERVER-876) Add with-puppet-conf-files macro to jruby-testutils
It is occasionally desirable to only have a single test have the
with-puppet-conf behavior of copying files into a directory, running a
test, and then cleaning the files up. This was previously impossible,
so this commit adds a macro that can be used in a test to accomplish
this. Here a with-puppet-conf-files macro is added, and the
with-puppet-conf-fixture is updated to use the macro under the hood.