(SERVER-876) Add test for current-code-id function
A previous commit moved where the current-code-id function was called
during request handling into the jruby request handler. In order to
ensure that current-code-id is always calculated with the same jruby
that is handling the request, this commit adds a test that makes a
request and inspects the state of the jruby pool while code id is being
calculated during the request.