(SERVER-1110) Introduce environment_classes endpoint
This commit introduces a new HTTP endpoint, environment_classes, which
returns information about the available classes on a per environment
basis, including the names of the classes, associated parameter names
and types, and the default values for any parameters, where applicable.

This commit also moves some of the JRuby request handling helper
functions from the request-handler-core namespace into a new namespace,
jruby-request, so that the backend implementation of the new
environment_classes endpoint can reuse some of those same functions
without having to be processed by the request-handler-service.