(MAINT) Bump jruby-pool-code-id test timeout
This commit bumps up the maximum amount of time that the
`test-jruby-pool-not-full-during-code-id-generation` test will wait for
a catalog request to complete from 5 seconds to 10 seconds.  This test
frequently failed for me locally when run in isolation from the command
line because the first catalog request takes right around 5 seconds,
although subsequent ones were much faster (~100 milliseconds).