(SERVER-1111) Generate and add etag to the environment_classes response
This commit generates a hash value from the JSON response to an
environment_classes request and adds that hash to the HTTP response in
an Etag header.  The commit includes some work to resort maps in the
response body prior to JSON serialization so that the generated hash is
consistent even if the order in which keys had been inserted into the
original sub-maps may differ from one object creation to another.

To avoid having to redo the sorting-related logic later when the changes
land for SERVER-1126, this commit also handles the case where the
underlying classes_per_environment call might return an error in the
class info when failing to parse a file.