(SERVER-1111) Add cache of environment class info etags
This commit adds a cache of the latest retrieved etags for
environment_class API queries to the jruby puppet service.  Calls to the
environment-cache endpoint, via the mark-all-environment(s)-expired
calls, invalidate entries in the cache.  This commit also includes some
refactoring of the environment_classes middleware functions.

This commit also upgrades the dependency on puppetlabs/http-client to
0.5.0.  This was needed in order for test functions to properly handle
an empty response body, which is delivered in cases where the
environment_classes endpoint returns an HTTP 304 (Not Modified) in
conjunction with an empty response body per the HTTP spec.