(SERVER-1111) Restrict class info map key coercion for RubySymbol
This commit changes the logic involved in sorting map keys for use in
calculating an environment class info etag to only coerce keys to
keywords for the types that `keyword` already handles - just string and
keyword - and `RubySymbol`.  This removes the prior logic which would
use `.toString` on the object as a coercion fallback, replacing that
instead with throwing an `IllegalArgumentException`.