(maint) Return 400 is no environment is in a catalog request
Previously, because with-code-id simply threw an exception, catalog
requests without an environment would return a 500. A bad request of 400
seems like a better match for the expected return code (as the client
controls whether or not an environment is included in the request), so
this commit updates with-code-id to throw a bad request if no
environment is present in the request.