(SERVER-1107) Throw if code-id command fails
Previously, we would only throw an exception if the execution
of the code-id command returned a non-zero exit code. This had
the effect of fast-failing a catalog request, but if the
execution of the command failed and an exception was thrown,
the catalog request would succeed with `nil` added to the catalog
as the code-id.

This commit changes the behavior of the execution of the code-id
command to throw an `IllegalStateException` any time we catch
an exception during code-id execution, ensuring that catalog
requests will always fail if there is an error generating the
code-id.