(SERVER-1107) Fail on non-zero exit from code-id-command
Previously nil was returned when the code-id-command exited non-zero.
This was undesirable so this commit simply throws an exception in the
case of a non-zero exit. The throw triggers a 500, which will fail the
catalog request.
Other exceptions that were previously being caught and rethrown in
execute-code-id-script! are now not being caught, so they should
similarly cause a 500 in a catalog request.