(SERVER-1153) Handle superfluous --gzip if-none-match header characters
This commit strips the characters "--gzip" from the end of an
'If-None-Match' HTTP header value when comparing the value to that of
the last etag stored for the environment_classes cache.  Previously, no
special compensation for these characters was being done.  This had led
to failures matching up the supplied 'If-None-Match' header value to
that of the last stored value, in cases where the Jetty web server was
appending the "--gzip" characters to the etag value as it delivered a
gzipped response body to a prior environment_classes request.