(SERVER-1130) Add a per-environment timestamp into the env-classes cache
This commit adds a "last-updated" timestamp per environment to the
environment classes cache.  The timestamp is used by incoming
environment_classes cache requests to determine if the cache info has
changed between the time the request starts parsing for info and when
the cache is about to be updated, e.g., if a call from a separate
request to mark-environment-expired were to occur during parsing for
class info in a request.  If the timestamp has been updated during that
window, the cache is not updated with the latest parsed data, since that
data may already be stale.

This commit also adds an integration test,
class-info-updated-after-cache-flush-during-prior-request, which
performs an environment cache flush while a class info request is in
progress and validates that the following class info request gets the
latest data.