(SERVER-1142) Sanely handle auth and param parsing for clojure routes
Previously clojure routes were unauthenticated if legacy auth was
enabled because it was presumed that all routes would be handed down
into puppet for auth via auth.conf. As that is not the case, this commit
separates the jruby request handler from a new clojure request wrapper.
The jruby request handler is applied to all ruby routes, and conditional
authentication is applied to that handler. The new clojure wrapper
unconditionally applies authentication and parses the request for
params.
The legacy routes service also needed a corresponding change to account
for the changes to get-wrapped-handler and root-routes.
Additionally, this commit defines schemas for some of the existing and
all of the new functions in master-core.