(SERVER-1142) Wrap entire clojure subtree with clojure-handler
Previously each clojure handler was wrapped with the
clojure-request-wrapper. This had the unfortunate consequence that we
could easily add a new clojure route in the future and forget to wrap it
with authentication or parameter parsing. This commit adds a call to
wrap-routes to wrap the entire clojure subtree with both authentication
and parameter parsing to avoid that.