(PDB-2329) Shared storage for terminii
In order to support high availability, the PuppetDB terminus needs
to implement sticky failover. That is, after failing over to another
PuppetDB instance, it should keep using that instance for future
requests as long as possible. This is tricky when running in
PuppetServer, since it hosts multiple JRuby instances.

This adds a shared java.util.concurrent.ConcurrentHashMap instance as
the :shared-terminus-state entry of the jruby pool context. It is passed
to each JRuby instance when it is constructed and passed into the ruby
code via the puppet_server_config map (in the shared_terminus_state
key). Finally, the Puppet::Server::Config initialize method probes for
the existence of well-known class from the PuppetDB
terminus (Puppet::Util::Puppetdb::SharedTerminusState). If this class
exists, the map is injected into via its shared_terminus_state property.

In a more perfect world, there would be a uniform terminus API by which
this map could be passed from puppet-server to the terminus.