(SERVER-1007) Print Timezone format with CA correctly on Java 8u60+
The version of Joda Time used by the previous version of clj-time had
an issue on Java 8u60+ that caused it to improperly print a different
timestamp.

This commit bumps the version of clj-time, which pins to a newer
version of Joda Time, which does not have this bug on Java 8u60 and
above.