(SERVER-1130) Mark expired between get and set in env-class-info tests
This commit changes the environment-class-info-tags test to mark an
environment tag as expired in between a get and corresponding set to
change the tag.  This is a more "real-world" test than the previous
version, which had instead just faked tag invalidation by providing a
different timestamp for the set than had been obtained for the prior
get.

This commit also changes an assertion to validate the content of an etag
rather than the response that etag was delivered in, consistent with the
corresponding failure string that was already in the assertion.