(SERVER-1148) Added timed waits in jruby-pool-int-tests
This commit adds some upper bounds on the maximum number of iterations /
wait time for specific operations to complete while running the
jruby-pool-int-tests.  The bounded operations include the maximum wait
time for a new pool to be initialized, instances to be borrowed, pool
flushes to complete, and for the puppetserver service stack to be
stopped.