(SERVER-1183) Add entry to env class cache if not present during get
This commit adds some logic behind the
`get-environment-class-info-tag-last-updated` call in the
jruby-puppet-service to update the environment class info cache with a
new entry for the environment being queried in the event that no prior
entry existed for that environment.  This allows any subsequent
`mark-all-environments-expired!` call that may be made to the service to
update the `last-updated` date in the cache for the environment and,
therefore, for the logic added in SERVER-1130 to avoid updating the
cache with a stale tag to be used for the case where the get precedes a
`mark-all-environments-expired!` call.