(SERVER-1183) Add exclamation point to get-last-updated function
This commit adds an exclamation point to the end of the
`get-environment-class-info-tag-last-updated!` function in the
jruby-puppet-service.  This was done in order to signify that a call to
the function could now have the side effect of a new environment entry
being added into the cache if it had not already existed when the call
was made.