(SERVER-1148) Bump test-restart-comes-back timeout to 5 minutes
This commit bumps the timeout that the `test-restart-comes-back` test
uses to wait for a restart to complete from 90 seconds to 5 minutes.  It
appears that 90 seconds may not have been a long enough time to wait
based on the results of some test runs in Travis, where the JRubyPuppet
pool instance initialization after the restart sometimes appears to not
have completed before the timeout is reached.