(SERVER-1185) Add enable setting for environment class cache
This commit adds a Trapperkeeper-level configuration setting for use in
enabling/disabling the environment class cache.  When set to "false"
(disabled), an Etag is no longer computed/returned with the payload for
a environment_classes API GET request.  When set to "true" (enabled),
the Etag will be computed and returned with the GET request payload as
was true in prior commits.