(SERVER-1185) Move environment-class-cache-enabled setting to jruby-puppet
Previously, the `environment-class-cache-enabled` setting was in a
`puppetserver` block. This was the only setting in this block. Ultimately, we
would like to separate out the `jruby-puppet` setting block to only contain
the jruby-specific settings and move many of the settings to a `puppetserver`
block. However, we are not at the point of doing that yet, and in the meantime
only having one setting in the `puppetserver` block seems confusing to explain
to users (why is this here when all the other seemingly-puppet server related
settings aren't here?). This commit moves this setting to the `jruby-puppet`
block for now.