(SERVER-1185) Remove environment-class-cache-enabled from schema
As `environment-class-cache-enabled` is not actually used by the jruby-puppet
service anywhere and doesn't need to be passed around everywhere, this commit
removes it from the JRubyPuppetConfig schema and `dissoc`'s it from the config
before checking it against this schema. In an ideal world we would have a
schema for what is allowed in the config file and another schema for the
things we pass around that are actually related to jruby, but that is an
improvement we will make in the future (SERVER-1189).

Also puts the setting in the right place in the example config file.