(SERVER-1160) Update execution.rb for complex commands
Previously, Puppet Server failed to correctly parse more complex commands
(with lots of string escaping) passed to it by Puppet for execution.

This commit fixes that problem by updating execution.rb such that when an
array is passed it no longer concatenates that array to be a string, but
instead the first argument is passed to `ShellUtils/executeCommand` as a
string and the rest are passed as an array.