(SERVER-1183) Removed unchecked protection for cache id increment
This commit removes protection for buffer overflow for increments to
the `cache-generation-id` given that it would be extremely unlikely to
ever practically hit the upper bound on the id before overflowing.

This commit also changes the function name
`updated-environment-class-info-entry` to
`update-environment-class-info-entry` so that a verb is used in the
name for more similarity with the names of other functions.