(SERVER-1194) Validate code-id, env for external scripts
This commit adds validation for inputs to the code id and code content
scripts. While ShellUtils ought to prevent any bash injection attacks,
we still deemed it wise to do validation on this sort of input.

To that end, this commit adds schemas for environment and code-id
strings and uses them at both the HTTP level and the schema level.

An old helper, `valid-code-id?`, was removed in favor of using schema.

Conflicts:
	test/integration/puppetlabs/general_puppet/general_puppet_int_test.clj