(SERVER-1196) Add time out to the class-info-cache-flush test
This commit adds a timeout to the initial wait that the
`class-info-updated-after-cache-flush-during-prior-request` does to see
that the getClassInfoForEnvironment method in a JRubyPuppet instance has
been invoked.  The current code waits indefinitely for that action,
which may never happen in the case that some higher-level failure has
occurred, e.g., the client fails to connect to the server, in which case
the test could hang.