(SERVER-1204) Update execution stub to correctly handle empty args
Previously, puppet server's execution stub assumed that if it was handed
an array, that the first element would be the command. This turns out
not to be true for all of puppet's uses of the execute method.
Specifically, config_version is often a script with arguments that is
handed as a single string in an array to Puppet::Util::Execution.execute
and then on to puppet server's execution stub.
This is a problem because something like '/bin/echo a b c' is not a valid
command. This commit addresses the problem by using the correct call to
ShellUtils in the case where there is a single element array. The
ShellUtils call will in that case correctly parse '/bin/echo a b c' into
a call to /bin/echo with a b and c as arguments.