(ORCH-1235) Add code_id to environment catalog requests
Similar to agent catalog requests, we now add code id for the given
environment to environment catalog requests.

This extends the logic for determining the request environment to now
check both the :environment URL segment as well as the "environment"
query param. If environment is part of the URL, we favor that over a
conflicting query param.