(SERVER-1215) Fix inconsistent logback versions
When we introduced the dependency on logstash-logback-encoder,
we accidentally started bringing in a newer version of some
logback artifacts than those that are brought in by tk and tk-j9.

This commit temporarily makes all of the logback dependency
versions explicit in the puppet server project, so that we can
get a quick bugfix release out.

Longer term we probably want to move all of these to TK directly,
since that is the repo that really handles all of our logging
logic.