(SERVER-864) Allow special characters in certificate_statuses key
Prior to this the certificate_statuses endpoint would not match calls
when the required key path segment contained anyting that wasn't an
alphanum, hyphen, underscore, or period. eg.
  "puppet-ca/v1/certificate_statuses/*"
would fail to match the specification of:
  ":ca-mount-endpoint/:version/certificate_statuses/:key"
even though there is no further requirements on the key path segment.

This behavior is determined by how we describe our path segment captures
to a third party library (bidi).

This patch updates our route description to allow the key path segment
to match any single path segment of one or more characters in length.