(SERVER-1151) Add the normalize request uri middleware to handlers
This commit adds `normalize-request-uri` to all of the handlers that
Puppet Server registers with the webserver service.  This enables the
value seen by the handlers for calls to `.getRequestUri` at the web
server layer to receive a form of the URI path component that has been
"normalized" - percent-decoded and with any requests which include
relative paths being rejected with an HTTP 400 (Bad Request).