(maint) Return 400 for incomplete URL to certificate_statuses
Previously, we were falling through to the default handler of our
web-routes if the certificate_statuses endpoint did not include its
required -- but ignored -- trailing key segment (which returns a 404).

This provides a default handler for the certificate_statuses endpoint
that will catch malformed URLs and return a 400 instead of a 404.

It also extracts providing a generic 400 response into a method that
is reused where appropriate.