(SERVER-1151) Update to TK 1.3.1 and remove logback version hack
This commit bumps us to the latest release of TK, which includes
explicit versioning for all of our logback artifacts to ensure
that we don't end up shipping incompatible versions of the various
logback components.  The commit also removes the previous hack
that we'd put in to temporarily fix the logback versioning issue.