(SERVER-977) refactor `delete-certificate!` & `process-csr-deletion!`
Prior to this `process-csr-deletion!` deleted CSRs and
`delete-certificate!` deleted certs and CSRs.

This removes CSR deletion from `delete-certificate!`, refactors
`process-csr-deletion!` (since its a new method for the containing PR)
to match `delete-certificate!` in name and arg position, and updates
code that called `delete-certificate!` to call it and the modified
version of `process-csr-deletion!`.