(PUP-5482) Toggle always_retry_plugins when available
A performance improvement was just merged into puppet. It is guarded by
a setting called `always_retry_plugins`, which defaults to true. To take
advantage of this improvement, we should set it to false. This commit
makes that change, but checks to see if the setting is available yet by
calling into `Puppet.settings.setting`.