(SERVER-1187) Adjust log levels for autosign actions
This commit makes two changes to improve the logging for autosigning.
First, if the autosign command generates error output or a non-zero
exit-code, a warning is generated in the logs. Second, it moves the
signing log message to info from debug. Previously this information
would only be available when logging at the debug level, which would
generate a substantial amount of noise.