(SERVER-1151) Bump keylength for double encoded request test to 1024
This commit bumps the keylength used in the generation of a certificate
by the double-encoded-request-not-allowed test from 512 bits to 1024
bits.  512 bits is below the minimum configured in the JDK java.security
default for our CI machines, causing the tests to fail when run in CI.

This commit also cleans up a bit of redundant code in the setup of the
SSL context for the double-encoded-request-not-allowed test.