(MAINT) Remove csr attrs and oid error logging
Previously, an error was logged if the configured CSR Attributes
file path or the configured Custom Trusted OID Mappings file
path failed an fs/file? check. However, Puppet has a default
value for both of these settings, meaning the error would be
logged on startup unless the user specifically adds those
configuration files.

This removes the logging of that error.