(SERVER-1187) Only log stderr once when autosigning
The previous commit updated the execute-autosign-command! function to
add a warning if stderr was generated which resulted in the stderr for
the autosign script being logged twice, once at debug and once at warn
levels. This commit adjusts the logging further to log the stdout
from the process at debug, and only log stderr at warn if there is any.