(SERVER-1245) Add trusted oid mapping to env classes int test
This commit adds a `trusted-oid-mapping-file` setting value that the
mock pupppetserver config service in the test uses.  This was added in
order for the newly added code which handles the trusted oid mapping
file to work without any problems - although it isn't directly related
to anything that the specific test validates.