(SERVER-1151) Added catalog name comparison integration assertions
This commit points the Ruby Puppet submodule at a commit in the Puppet
CVE repo and adds a couple of assertions integration assertions which
validate that the changes from PUP-6086 allow a catalog to be requested
with only a single pass of percent-decoding of the request path being
done, before authorization.  Prior to the changes from PUP-6086, these
assertions would have failed because the request path would have been
decoded twice - once before authorization and once in the indirector
layer in Ruby Puppet code.