(SERVER-1206) Restore dash in puppet-server for webserver server config
This commit restores the use of a dash in puppet-server in the handling
of the webserver config's server-id.  A previous commit had removed it
but this commit restores it out of fear that removing it may cause
compatibility issues for PE users that had already become dependent upon
the presence of the dash character.