(SERVER-1247) Packaging changes for bootstrap improvements
This commit updates ezbake files and settings to take advantage of
ezbake and tk improvements to allow a project to split its
bootstraps into two services.d directories.

To accommodate a new directory of files to be placed under /opt/,
the ezbake/config dir now has two subdirectories, public and private.
Everything under public is the same as it was before, under config/,
with the exception of the new services.d dir, which contains ca.cfg.

The remainder of the bootstrap entries are in
ezbake/config/private/services.d/bootstrap.cfg

The files under /ext/config have been moved to /ext/config/public
to mirror the ezbake folder structure