(maint) Bump puppet and hiera submodules
For some reason (aka I used `git commit -a` thinking I was committing one
change to project.clj but not realizing I hadn't done a submodule update - see
https://github.com/puppetlabs/puppetserver/pull/1093/commits/636b1a953dfb83c698fed09d86aaac7c281acba3)
when stable was merged up to master, the bumping of puppet and hiera
submodules didn't get carried forward, but the PUPPET_BUILD_VERSION in the
acceptance helper was bumped to 1.5.0. This bumps the puppet and hiera
submodules to the versions that will ship in puppet-agent 1.5.0 (and what is
currently on stable).