(SERVER-1163) Remove jruby code and tests
This commit removes all of the code that was extracted into the jruby-utils
library, as well as tests for that code.

puppetserver code was then tied back up to the new jruby-utils code. This is
where the majority of the effort has gone, since many small things changed.
Most of the changes are in the jruby-puppet-service and related code.

Some important changes are:
* The jruby pool is no longer puppetserver specific, so the code has been made
  more general to accommodate this
* The work of initializing the jruby-pool is split between puppetserver and
  jruby-utils. There are now lifecycle functions passed to the
  pool-manager-service that are run automatically to finish the initialization.
  These setup the jruby scripting container, the pool instances, as well as
  handle cleanup and shutdown.
* Calls to jruby-puppet namespaces were replaces with jruby-utils calls
* A new pool-manager service from jruby-utils is used to create the
  pool-context in the jruby-puppet-service
* the with-jruby-puppet macro has been replaced by a more general macro from