Modify the pre-suite to ignore the PUPPET_VERSION/puppet_version envvar/beaker setting and use PUPPET_BUILD_VERSION/puppet_build_version envvar/beaker setting consistently.
In 30_install_dev_repos.rb, we install a puppet-agent development
repo on the master based on the value of **puppet_build_version**.
Later, in 70_install_puppet.rb, we _attempt_ to install the agent,
but the logic is precisely inverted from what we want, so we end
up pulling in unwanted
agent code.

At line 39, we fetch and check the value of **puppet_version**; if
it is set, we fall through to the **elsif** on line 46 and install
the default puppet-agent package, which is the _development_ version
in the repo installed in 30_install_dev_repos.rb. But, if it not
set, we fall through to line 49, where we fetch the value again (?)
this time _without_ checking it's value, and then attempt to install
the agent using its value, which is empty. As a result, we again
install the default puppet-agent package available, which is the
dev version installed earlier.

PUPPET_VERSION is not used in CI, AFAICT, but PUPPET_BUILD_VERSION
is, so this change modifies the pre-suite to use the value of
PUPPET_BUILD_VERSION.