(maint) use thrown+? instead of custom helper
This was a pretty straightforward change except for where we were
looping over exceptions and checking thrown-with-slingshot?; the
thrown+? macro assumes, when it sees a name passed to it (instead of a
sequence literal) that that name is a function. I had to pass functions
instead of more succinct slingshot selectors.

This is why we can't have nice things.