(maint) Add semantic_puppet gem
Note that puppet is planning to add semantic_puppet as a
hard dependency. This commit adds the semantic_puppet gem
as a git submodule under ruby/, so that developers can
work with versions of puppet that have that dependency.

It also updates the ruby load path in several places to
add semantic_puppet.