(PUP-6335) Add note explaining duplication in certificate_authority.clj
The information captured in various vars in certificate_authority.clj is
duplicated in the puppet codebase. Creating a shared oids file isn't
easily achieveable today, so instead this commit adds a comment above
those vars in certificate_authority.clj explaining that any changes to
the oids need to be replicated in the puppet codebase.