(SERVER-1440) Close up test-503-when-app-shuts-down race condition a bit
This commit adds an initial GET request to the environments endpoint to
the `test-503-when-app-shuts-down` test.  Without this extra call, it
is more likely for the environment ping logic in the test to exhaust up to
the maximum configured 10 second wait if the initialization of the second
JRubyPuppet instance, the one servicing the environments requests, takes
a sufficiently long amount of time.  This would, in turn, cause
assertions related to the service stop to fail because the stop was not
far enough along before the assertions were made.