(SERVER-1454) Bump puppet pins to pick up json-pure pin
This commit bumps up the Puppet submodule and puppet package versions in
order to pick up a pin of the json-pure gem to '~> 1.8', which allows
the Ruby spec tests to continue to be run under Ruby 1.9.3.  Previously,
the version of the json-pure gem that was being installed was not
pinned.  The recently released version, 2.0.2, dropped support for
running under versions of Ruby earlier than 2.0.