(MAINT) Update auth_rules file_bucket test to expect 400 error
This commit updates the test in default_rules which allows an authorized
file_bucket_file request to expect an HTTP status code of 400 (Bad
Request) instead of 500 (Server Error).  Through work related to
PUP-3827 which was done in Ruby Puppet code, 400 (Bad Request) is now
the proper response that the test should see when including a malformed
md5 value.