(SERVER-1456) Upgrade nss before puppetserver install in CI pre-suite
This commit ensures that the nss package is upgraded to the latest
available version before openjdk and puppetserver are installed during a
CI test run.  Prior to this change, the nss package was not upgraded,
which could lead to puppetserver erroring out with a
`java.lang.InternalError` at service startup - due to apparent version
conflicts between pre-3.21.0 versions of nss and openjdk 1.7.0_111 and
later.