(maint) Do not fail if journalctl is empty
This commit allows the `journal -u` command to return with a failure.
This would happen if a service has no output for some reason. Although
this is not common, this is the case for puppetserver on sles 12. The
start and stop of this process is not recorded in the systemd journal.
As a result, `journalctl -u puppetserver` will fail even though the
puppetserver service is running successfully.